define serializeable protobuf for test results
