+++ compiler/damlc/BUILD.bazel
@@ -225,6 +225,7 @@ da_haskell_library(
         "//compiler/repl-service/client",
         "//compiler/scenario-service/client",
         "//compiler/scenario-service/protos:scenario_service_haskell_proto",
+        "//compiler/scenario-service/protos:test_results_haskell_proto",
         "//daml-assistant:daml-project-config",
         "//daml-assistant/daml-helper:daml-helper-lib",
         "//daml-lf/archive:daml_lf_dev_archive_haskell_proto",
