+++ compiler/damlc/lib/DA/Cli/Damlc/Test.hs
@@ -25,13 +25,10 @@ import DA.Daml.Options.Types
 import DA.Daml.Project.Consts (sdkPathEnvVar)
 import qualified DA.Pretty
 import qualified DA.Pretty as Pretty
-import Data.Either
 import Data.Foldable (fold)
 import qualified Data.HashSet as HashSet
 import Data.List.Extra
-import qualified Data.Map.Strict as M
 import Data.Maybe
-import qualified Data.NameMap as NM
 import qualified Data.Text as T
 import qualified Data.Text.IO as TIO
 import qualified Data.Text.Lazy as TL
@@ -46,6 +43,7 @@ import Development.IDE.Types.Location
 import qualified Development.Shake as Shake
 import Safe
 import qualified ScenarioService as SS
+import qualified DA.Cli.Damlc.Test.TestResults as TR
 import System.Console.ANSI (SGR(..), setSGRCode, Underlining(..), ConsoleIntensity(..))
 import System.Directory (createDirectoryIfMissing)
 import System.Environment.Blank
@@ -56,8 +54,6 @@ import System.IO.Error (isPermissionError, isAlreadyExistsError, isDoesNotExistE
 import qualified Text.XML.Light as XML
 import qualified Text.Blaze.Html.Renderer.Text as Blaze
 import qualified Text.Blaze.Html4.Strict as Blaze
-import Text.Printf
-
 
 newtype UseColor = UseColor {getUseColor :: Bool}
 newtype ShowCoverage = ShowCoverage {getShowCoverage :: Bool}
@@ -73,23 +69,6 @@ execTest inFiles runAllTests coverage color mbJUnitOutput opts tableOutputPath t
         testRun h inFiles (optDamlLfVersion opts) runAllTests coverage color mbJUnitOutput tableOutputPath transactionsOutputPath
         diags <- getDiagnostics h
         when (any (\(_, _, diag) -> Just DsError == _severity diag) diags) exitFailure
-
-data LocalOrExternal
-    = Local LF.Module
-    | External LF.ExternalPackage
-    deriving (Show, Eq)
-
-isLocal :: LocalOrExternal -> Bool
-isLocal (Local _) = True
-isLocal _ = False
-
-loeGetModules :: LocalOrExternal -> [(LF.Module, a -> LF.Qualified a)]
-loeGetModules (Local mod) = pure (mod, LF.Qualified LF.PRSelf (LF.moduleName mod))
-loeGetModules (External pkg) =
-    [ (mod, qualifier)
-    | mod <- NM.elems $ LF.packageModules $ LF.extPackagePkg pkg
-    , let qualifier = LF.Qualified (LF.PRImport (LF.extPackageId pkg)) (LF.moduleName mod)
-    ]
 
 testRun ::
        IdeState
@@ -136,21 +115,21 @@ testRun h inFiles lfVersion (RunAllTests runAllTests) coverage color mbJUnitOutp
         else pure []
 
     let -- All Packages / Modules mentioned somehow
-        allPackages :: [LocalOrExternal]
-        allPackages = [Local mod | (_, _, mod, _) <- results] ++ map External extPkgs
+        allPackages :: [TR.LocalOrExternal]
+        allPackages = [TR.Local mod | (_, _, mod, _) <- results] ++ map TR.External extPkgs
 
         -- All results: subset of packages / modules that actually got scenarios run
-        allResults :: [(LocalOrExternal, [(VirtualResource, Either SSC.Error SS.ScenarioResult)])]
+        allResults :: [(TR.LocalOrExternal, [(VirtualResource, Either SSC.Error SS.ScenarioResult)])]
         allResults =
-            [(Local mod, result) | (_world, _file, mod, Just result) <- results]
-            ++ [(External pkg, result) | (pkg, Just result) <- extResults]
+            [(TR.Local mod, result) | (_world, _file, mod, Just result) <- results]
+            ++ [(TR.External pkg, result) | (pkg, Just result) <- extResults]
 
     -- print test summary after all tests have run
     printSummary color (concatMap snd allResults)
 
     -- print total test coverage
-    printTestCoverage
-        coverage
+    TR.printTestCoverage
+        (getShowCoverage coverage)
         allPackages
         allResults
 
@@ -277,306 +256,6 @@ printSummary color res =
         , "Test Summary" <> setSGRCode []
         ]
     printScenarioResults color res
-
-data ContractIdentifier = ContractIdentifier
-    { package :: Maybe T.Text -- `package == Nothing` means local package
-    , qualifiedName :: T.Text
-    }
-    deriving (Eq, Ord, Show)
-
-data ChoiceIdentifier = ChoiceIdentifier
-    { packageContract :: ContractIdentifier
-    , choice :: T.Text
-    }
-    deriving (Eq, Ord, Show)
-
-lfTemplateIdentifier :: LF.Qualified LF.Template -> ContractIdentifier
-lfTemplateIdentifier = lfMkNameIdentifier . fmap LF.tplTypeCon
-
-lfInterfaceIdentifier :: LF.Qualified LF.DefInterface -> ContractIdentifier
-lfInterfaceIdentifier = lfMkNameIdentifier . fmap LF.intName
-
-lfMkNameIdentifier :: LF.Qualified LF.TypeConName -> ContractIdentifier
-lfMkNameIdentifier LF.Qualified { qualPackage, qualModule, qualObject } =
-    let package =
-            case qualPackage of
-              LF.PRSelf -> Nothing
-              LF.PRImport (LF.PackageId pid) -> Just pid
-        qualifiedName =
-            LF.moduleNameString qualModule
-                <> ":"
-                <> T.concat (LF.unTypeConName qualObject)
-    in
-    ContractIdentifier { package, qualifiedName }
-
-ssIdentifierToIdentifier :: SS.Identifier -> ContractIdentifier
-ssIdentifierToIdentifier SS.Identifier {SS.identifierPackage, SS.identifierName} =
-    let package = do
-            pIdSumM <- identifierPackage
-            pIdSum <- SS.packageIdentifierSum pIdSumM
-            case pIdSum of
-                SS.PackageIdentifierSumSelf _ -> Nothing
-                SS.PackageIdentifierSumPackageId pId -> Just $ TL.toStrict pId
-        qualifiedName = TL.toStrict identifierName
-    in
-    ContractIdentifier { package, qualifiedName }
-
-printTestCoverage ::
-    ShowCoverage
-    -> [LocalOrExternal]
-    -> [(LocalOrExternal, [(VirtualResource, Either SSC.Error SS.ScenarioResult)])]
-    -> IO ()
-printTestCoverage ShowCoverage{getShowCoverage} allPackages results
-  | any (isLeft . snd) $ concatMap snd results = pure ()
-  | otherwise = printReport
-  where
-    printReport :: IO ()
-    printReport =
-        let countWhere pred = M.size . M.filter pred
-            pctage :: Int -> Int -> Double
-            pctage _ 0 = 100
-            pctage n d = max 0 $ min 100 $ 100 * fromIntegral n / fromIntegral d
-
-            allTemplates = templatesDefinedIn allPackages
-            localTemplates = M.filterWithKey pred allTemplates
-              where
-                pred (ContractIdentifier Nothing _) _ = True
-                pred _ _ = False
-            localTemplatesCreated = M.intersection allCreatedContracts localTemplates
-
-            allTemplateChoices = templateChoicesDefinedIn allPackages
-            localTemplateChoices = M.filterWithKey pred allTemplateChoices
-              where
-                pred (ChoiceIdentifier (ContractIdentifier Nothing _) _) _ = True
-                pred _ _ = False
-            localTemplateChoicesExercised = M.intersection allExercisedChoices localTemplateChoices
-
-            allInterfaces = interfacesDefinedIn allPackages
-            allImplementations = interfaceImplementationsDefinedIn allPackages
-            fillInImplementation (ifaceId, _) (loe, instanceBody) = (loe, instanceBody, M.lookup ifaceId allInterfaces)
-
-            allImplementationChoices = M.fromList $ do
-                (k@(_, contractId), (loe, body, mdef)) <- M.toList $ M.mapWithKey fillInImplementation allImplementations
-                def <- maybeToList mdef
-                choice <- NM.toList $ LF.intChoices $ LF.qualObject def
-                let name = LF.unChoiceName $ LF.chcName choice
-                guard (name /= "Archive")
-                pure (ChoiceIdentifier contractId name, (k, loe, body, def, choice))
-
-            localImplementationChoices = M.filter pred allImplementationChoices
-              where
-                pred (_, loe, _, _, _) = isLocal loe
-            localImplementationChoicesExercised = M.intersection allExercisedChoices localImplementationChoices
-            externalImplementationChoices = M.filter pred allImplementationChoices
-              where
-                pred (_, loe, _, _, _) = not (isLocal loe)
-            externalImplementationChoicesExercised = M.intersection allExercisedChoices externalImplementationChoices
-
-            externalTemplates = M.filterWithKey pred allTemplates
-              where
-                pred (ContractIdentifier (Just _) _) _ = True
-                pred _ _ = False
-            externalTemplatesCreated = M.intersection allCreatedContracts externalTemplates
-
-            externalTemplateChoices = M.filterWithKey pred allTemplateChoices
-              where
-                pred (ChoiceIdentifier (ContractIdentifier (Just _) _) _) _ = True
-                pred _ _ = False
-            externalTemplateChoicesExercised = M.intersection allExercisedChoices externalTemplateChoices
-
-            showCoverageReport :: (k -> String) -> String -> M.Map k a -> [String]
-            showCoverageReport printer variety names
-              | not getShowCoverage = []
-              | otherwise =
-                [ printf "  %s: %d" variety (M.size names)
-                ] ++ [ "    " ++ printer id | id <- M.keys names ]
-        in
-        putStrLn $
-        unlines $
-        concat
-        [ [ printf "Modules internal to this package:" ]
-        -- Can't have any external tests that exercise internals, as that would
-        -- require a circular dependency, so we only report local test results
-        , let defined = M.size localTemplates
-              created = M.size localTemplatesCreated
-              neverCreated = M.difference localTemplates localTemplatesCreated
-          in
-          [ printf "- Internal templates"
-          , printf "  %d defined" defined
-          , printf "  %d (%5.1f%%) created" created (pctage created defined)
-          ] ++ showCoverageReport printContractIdentifier "internal templates never created" neverCreated
-        , let defined = M.size localTemplateChoices
-              exercised = M.size localTemplateChoicesExercised
-              neverExercised = M.difference localTemplateChoices localTemplateChoicesExercised
-          in
-          [ printf "- Internal template choices"
-          , printf "  %d defined" defined
-          , printf "  %d (%5.1f%%) exercised" exercised (pctage exercised defined)
-          ] ++ showCoverageReport printChoiceIdentifier "internal template choices never exercised" neverExercised
-        , let localImplementations = M.filter (isLocal . fst) allImplementations
-              defined = M.size localImplementations
-              (internal, external) = M.partitionWithKey (\(ifaceId, _) _ -> isNothing (package ifaceId)) localImplementations
-          in
-          [ printf "- Internal interface implementations"
-          , printf "  %d defined" defined
-          , printf "    %d internal interfaces" (M.size internal)
-          , printf "    %d external interfaces" (M.size external)
-          ]
-        , let defined = M.size localImplementationChoices
-              exercised = M.size localImplementationChoicesExercised
-              neverExercised = M.difference localImplementationChoices localImplementationChoicesExercised
-          in
-          [ printf "- Internal interface choices"
-          , printf "  %d defined" defined
-          , printf "  %d (%5.1f%%) exercised" exercised (pctage exercised defined)
-          ] ++ showCoverageReport printChoiceIdentifier "internal interface choices never exercised" neverExercised
-        , [ printf "Modules external to this package:" ]
-        -- Here, interface instances can only refer to external templates and
-        -- interfaces, so we only report external interface instances
-        , let defined = M.size externalTemplates
-              createdAny = M.size externalTemplatesCreated
-              createdInternal = countWhere (any isLocal) externalTemplatesCreated
-              createdExternal = countWhere (not . all isLocal) externalTemplatesCreated
-              neverCreated = M.difference externalTemplates externalTemplatesCreated
-          in
-          [ printf "- External templates"
-          , printf "  %d defined" defined
-          , printf "  %d (%5.1f%%) created in any tests" createdAny (pctage createdAny defined)
-          , printf "  %d (%5.1f%%) created in internal tests" createdInternal (pctage createdInternal defined)
-          , printf "  %d (%5.1f%%) created in external tests" createdExternal (pctage createdExternal defined)
-          ] ++ showCoverageReport printContractIdentifier "external templates never created" neverCreated
-        , let defined = M.size externalTemplateChoices
-              exercisedAny = M.size externalTemplateChoicesExercised
-              exercisedInternal = countWhere (any isLocal) externalTemplateChoicesExercised
-              exercisedExternal = countWhere (not . all isLocal) externalTemplateChoicesExercised
-              neverExercised = M.difference externalTemplateChoices externalTemplateChoicesExercised
-          in
-          [ printf "- External template choices"
-          , printf "  %d defined" defined
-          , printf "  %d (%5.1f%%) exercised in any tests" exercisedAny (pctage exercisedAny defined)
-          , printf "  %d (%5.1f%%) exercised in internal tests" exercisedInternal (pctage exercisedInternal defined)
-          , printf "  %d (%5.1f%%) exercised in external tests" exercisedExternal (pctage exercisedExternal defined)
-          ] ++ showCoverageReport printChoiceIdentifier "external template choices never exercised" neverExercised
-        , let defined = countWhere (not . isLocal . fst) allImplementations
-          in
-          [ printf "- External interface implementations"
-          , printf "  %d defined" defined
-          ]
-        , let defined = M.size externalImplementationChoices
-              exercisedAny = M.size externalImplementationChoicesExercised
-              exercisedInternal = countWhere (any isLocal) externalImplementationChoicesExercised
-              exercisedExternal = countWhere (not . all isLocal) externalImplementationChoicesExercised
-              neverExercised = M.difference externalImplementationChoices externalImplementationChoicesExercised
-          in
-          [ printf "- External interface choices"
-          , printf "  %d defined" defined
-          , printf "  %d (%5.1f%%) exercised in any tests" exercisedAny (pctage exercisedAny defined)
-          , printf "  %d (%5.1f%%) exercised in internal tests" exercisedInternal (pctage exercisedInternal defined)
-          , printf "  %d (%5.1f%%) exercised in external tests" exercisedExternal (pctage exercisedExternal defined)
-          ] ++ showCoverageReport printChoiceIdentifier "external interface choices never exercised" neverExercised
-        ]
-
-    templatesDefinedIn :: [LocalOrExternal] -> M.Map ContractIdentifier (LF.Qualified LF.Template)
-    templatesDefinedIn localOrExternals = M.fromList
-        [ (lfTemplateIdentifier templateInfo, templateInfo)
-        | localOrExternal <- localOrExternals
-        , (module_, qualifier) <- loeGetModules localOrExternal
-        , template <- NM.toList $ LF.moduleTemplates module_
-        , let templateInfo = qualifier template
-        ]
-
-    interfacesDefinedIn :: [LocalOrExternal] -> M.Map ContractIdentifier (LF.Qualified LF.DefInterface)
-    interfacesDefinedIn localOrExternals = M.fromList
-        [ (lfInterfaceIdentifier interfaceInfo, interfaceInfo)
-        | localOrExternal <- localOrExternals
-        , (module_, qualifier) <- loeGetModules localOrExternal
-        , interface <- NM.toList $ LF.moduleInterfaces module_
-        , let interfaceInfo = qualifier interface
-        ]
-
-    templateChoicesDefinedIn :: [LocalOrExternal] -> M.Map ChoiceIdentifier (LF.Qualified LF.Template, LF.TemplateChoice)
-    templateChoicesDefinedIn localOrExternals = M.fromList
-        [ (ChoiceIdentifier templateIdentifier name, (templateInfo, choice))
-        | (templateIdentifier, templateInfo) <- M.toList $ templatesDefinedIn localOrExternals
-        , choice <- NM.toList $ LF.tplChoices $ LF.qualObject templateInfo
-        , let name = LF.unChoiceName $ LF.chcName choice
-        ]
-
-    interfaceImplementationsDefinedIn :: [LocalOrExternal] -> M.Map (ContractIdentifier, ContractIdentifier) (LocalOrExternal, LF.InterfaceInstanceBody)
-    interfaceImplementationsDefinedIn localOrExternals = M.fromList $
-        [ ((lfMkNameIdentifier tpiInterface, templateIdentifier), (loe, tpiBody))
-        | loe <- localOrExternals
-        , (templateIdentifier, templateInfo) <- M.toList $ templatesDefinedIn [loe]
-        , LF.TemplateImplements { tpiInterface, tpiBody }
-            <- NM.toList $ LF.tplImplements $ LF.qualObject templateInfo
-        ] ++
-        [ ((interfaceIdentifier, lfMkNameIdentifier iciTemplate), (loe, iciBody))
-        | loe <- localOrExternals
-        , (interfaceIdentifier, interfaceInfo) <- M.toList $ interfacesDefinedIn [loe]
-        , LF.InterfaceCoImplements { iciTemplate, iciBody }
-            <- NM.toList $ LF.intCoImplements $ LF.qualObject interfaceInfo
-        ]
-
-    allCreatedContracts :: M.Map ContractIdentifier [LocalOrExternal]
-    allCreatedContracts = M.unionsWith (<>) $ map (uncurry createdContracts) results
-
-    allExercisedChoices :: M.Map ChoiceIdentifier [LocalOrExternal]
-    allExercisedChoices = M.unionsWith (<>) $ map (uncurry exercisedChoices) results
-
-    createdContracts :: LocalOrExternal -> [(VirtualResource, Either SSC.Error SS.ScenarioResult)] -> M.Map ContractIdentifier [LocalOrExternal]
-    createdContracts loe results =
-        M.fromList $
-        [ (ssIdentifierToIdentifier identifier, [loe])
-        | n <- scenarioNodes results
-        , Just (SS.NodeNodeCreate SS.Node_Create {SS.node_CreateContractInstance}) <-
-              [SS.nodeNode n]
-        , Just contractInstance <- [node_CreateContractInstance]
-        , Just identifier <- [SS.contractInstanceTemplateId contractInstance]
-        ]
-
-    exercisedChoices :: LocalOrExternal -> [(VirtualResource, Either SSC.Error SS.ScenarioResult)] -> M.Map ChoiceIdentifier [LocalOrExternal]
-    exercisedChoices loe results =
-        M.fromList $
-        [ (choiceIdentifier, [loe])
-        | n <- scenarioNodes results
-        , Just (SS.NodeNodeExercise SS.Node_Exercise { SS.node_ExerciseTemplateId
-                                                     , SS.node_ExerciseChoiceId
-                                                     }) <- [SS.nodeNode n]
-        , Just identifier <- [node_ExerciseTemplateId]
-        , let choiceIdentifier = ChoiceIdentifier (ssIdentifierToIdentifier identifier) (TL.toStrict node_ExerciseChoiceId)
-        ]
-
-    scenarioNodes :: [(VirtualResource, Either SSC.Error SS.ScenarioResult)] -> [SS.Node]
-    scenarioNodes results =
-        [ node
-        | (_virtualResource, Right result) <- results
-        , node <- V.toList $ SS.scenarioResultNodes result
-        ]
-
-    pkgIdToPkgName :: T.Text -> T.Text
-    pkgIdToPkgName targetPid =
-        case mapMaybe isTargetPackage allPackages of
-          [] -> targetPid
-          [matchingPkg] -> maybe targetPid (LF.unPackageName . LF.packageName) $ LF.packageMetadata $ LF.extPackagePkg matchingPkg
-          _ -> error ("pkgIdToPkgName: more than one package matching name " <> T.unpack targetPid)
-        where
-            isTargetPackage loe
-                | External pkg <- loe
-                , targetPid == LF.unPackageId (LF.extPackageId pkg)
-                = Just pkg
-                | otherwise
-                = Nothing
-
-    printContractIdentifier :: ContractIdentifier -> String
-    printContractIdentifier ContractIdentifier { package, qualifiedName } =
-        T.unpack $ maybe
-            qualifiedName
-            (\pId -> pkgIdToPkgName pId <> ":" <> qualifiedName)
-            package
-
-    printChoiceIdentifier :: ChoiceIdentifier -> String
-    printChoiceIdentifier ChoiceIdentifier { packageContract, choice } =
-        printContractIdentifier packageContract <> ":" <> T.unpack choice
 
 printScenarioResults :: UseColor -> [(VirtualResource, Either SSC.Error SS.ScenarioResult)] -> IO ()
 printScenarioResults color results = do
