+++ compiler/scenario-service/protos/BUILD.bazel
@@ -83,3 +83,45 @@ da_haskell_library(
     ],
     visibility = ["//visibility:public"],
 )
+
+# FIXME(JM): Modify compile-proto-file to allow specifying
+# the output file and module name rather than directory.
+genrule(
+    name = "test_results_haskell_proto_src",
+    srcs = ["test_results.proto", "scenario_service.proto", "//daml-lf/archive:daml_lf_1.dev_archive_proto_srcs"],
+    outs = ["TestResults.hs"],
+    cmd = """
+    ORIGPWD=$$PWD
+    cd compiler/scenario-service/protos
+    $$ORIGPWD/$(location @proto3-suite//:compile-proto-file) \
+      --includeDir ./                                        \
+      --includeDir $$ORIGPWD/daml-lf/archive/src/main/protobuf/        \
+      --proto test_results.proto --out "$$ORIGPWD/$(@D)"
+  """,
+    tools = [
+        "@proto3-suite//:compile-proto-file",
+    ],
+)
+
+da_haskell_library(
+    name = "test_results_haskell_proto",
+    srcs = [":test_results_haskell_proto_src"],
+    compiler_flags = ["-Wno-deriving-defaults"],
+    hackage_deps = [
+        "base",
+        "bytestring",
+        "containers",
+        "deepseq",
+        "grpc-haskell",
+        "grpc-haskell-core",
+        "proto3-suite",
+        "proto3-wire",
+        "text",
+        "vector",
+    ],
+    deps = [
+        "//daml-lf/archive:daml_lf_dev_archive_haskell_proto",
+        ":scenario_service_haskell_proto",
+    ],
+    visibility = ["//visibility:public"],
+)
